requirements.txt
clickhouse-driver==0.2.9
pandas==2.2.2
PyYAML==6.0.2
python-dateutil==2.9.0.post0
streamlit==1.35.0